+++ packages/physics/src/index.ts
@@ -1,3 +1,4 @@
 export * from './ohms'
 export * from './pendulum'
 export * from './lens'
+export * from './waveInterference'